changed REAL(8), COMPLEX(8) to REAL(DP), COMPLEX(DP) in util.F90 and localiz_rad.F90. Add functions description in fftpack2.F90

+++ code/source_f90/localize_rad.F90
@@ -1017,7 +1017,7 @@ REAL(DP) FUNCTION avermatrix(a,vec,ndim,kdimin)
 USE params, ONLY: DP
 IMPLICIT NONE
 
-!REAL(8) :: avermatrix
+!REAL(DP) :: avermatrix
 REAL(DP), INTENT(IN)                       :: a(kdimin,kdimin)
 REAL(DP), INTENT(IN)                       :: vec(kdimin)
 INTEGER, INTENT(IN)                      :: ndim
